remove default sanitization featues from design node conversion for figma
