+++ core/utils/index.ts
@@ -1,9 +1,11 @@
 export * from "./check-if-root";
+export * from "./check-if-auto-layout";
 export * from "./filter-fills";
 export * from "./has-image";
 export * from "../features/constraints/lcrs";
 export * from "./retrieve-fill";
 export * from "./retrieve-image-fills";
+export * from "./retrieve-colliding-children";
 export * from "./nodetype-map";
 export * from "./primary-color";
 export * from "./children";
